Updated directory and files structure in ball_e_gui
