+++ src/components/component_button.py
@@ -1,6 +1,14 @@
-from PyQt5.QtWidgets import QPushButton, QSizePolicy
+try:
+    import pathlib
+    import sys
+    sys.path.append(
+        "{}/Developer/ball_e_gui/src/helpers".format(pathlib.Path.home()))
 
-import style_constants as sc
+    import style_constants as sc
+except ImportError:
+    print("Imports failed")
+finally:
+    from PyQt5.QtWidgets import QPushButton, QSizePolicy
 
 
 class GenericButton(QPushButton):
