+++ src/screens/screen_help_calibration.py
@@ -1,34 +1,33 @@
-import sys
+try:
+    import pathlib
+    import sys
+    sys.path.append(
+        "{}/Developer/ball_e_gui/src/components".format(pathlib.Path.home()))
+    sys.path.append(
+        "{}/Developer/ball_e_gui/src/helpers".format(pathlib.Path.home()))
+    sys.path.append(
+        "{}/Developer/ball_e_gui/src/windows".format(pathlib.Path.home()))
 
-from PyQt5.QtCore import Qt
-from PyQt5.QtGui import QFont
-from PyQt5.QtWidgets import (QApplication, QListWidget, QListWidgetItem,
-                             QMessageBox, QVBoxLayout, QWidget)
-
-import helper_doc_reader as hdr
-from component_labels import ProfileLabel
-from component_modal import Modal
-from component_toolbar import ToolbarComponent
-from window_test import TestWindow
-
-
-class ListItem(QListWidgetItem):
-    def __init__(self, text):
-        super().__init__()
-        self.setText(text)
-        font = QFont()
-        font.setPixelSize(22)
-        self.setFont(font)
+    import helper_doc_reader as hdr
+    from component_labels import ProfileLabel
+    from component_list_item import ListItem
+    from component_modal import Modal
+    from component_toolbar import ToolbarComponent
+    from window_test import TestWindow
+except ImportError:
+    print("Imports failed")
+finally:
+    from PyQt5.QtWidgets import QApplication, QListWidget, QVBoxLayout, QWidget
 
 
-class TrainingHelpScreen(QWidget):
+class CalibrationHelpScreen(QWidget):
     def __init__(self, parent=None):
         super().__init__(parent=parent)
 
-        self.window_title = "Help: Training"
+        self.window_title = "Help: Calibration"
 
         # Get the docs for this section
-        doc_reader = hdr.DocReader('training/training_doc.md')
+        doc_reader = hdr.DocReader('calibration/calibration_doc.md')
         self.docs = doc_reader.get_doc()
 
         screen_layout = QVBoxLayout()
@@ -69,7 +68,7 @@ class TrainingHelpScreen(QWidget):
 
 def main():
     app = QApplication(sys.argv)
-    win = TestWindow(TrainingHelpScreen())
+    win = TestWindow(CalibrationHelpScreen())
     win.show()
     sys.exit(app.exec_())
