+++ src/screens/screen_profiles.py
@@ -1,12 +1,19 @@
-import sys
+try:
+    import pathlib
+    import sys
+    sys.path.append(
+        "{}/Developer/ball_e_gui/src/components".format(pathlib.Path.home()))
+    sys.path.append(
+        "{}/Developer/ball_e_gui/src/windows".format(pathlib.Path.home()))
 
-from PyQt5.QtCore import Qt
-from PyQt5.QtWidgets import (QApplication, QHBoxLayout, QLabel, QMainWindow,
-                             QPushButton, QSizePolicy, QVBoxLayout, QWidget)
+    from component_button import FullPageButton
+    from component_toolbar import ToolbarComponent
+    from window_test import TestWindow
 
-from component_button import FullPageButton, GenericButton
-from component_toolbar import ToolbarComponent
-from window_test import TestWindow
+except ImportError:
+    print("Imports failed")
+finally:
+    from PyQt5.QtWidgets import QApplication, QHBoxLayout, QVBoxLayout, QWidget
 
 
 class ProfilesScreen(QWidget):
