+++ src/screens/screen_training_goal_calibration_take_photo.py
@@ -1,18 +1,27 @@
-import sys
+try:
+    import pathlib
+    import sys
+    sys.path.append(
+        "{}/Developer/ball_e_gui/src/components".format(pathlib.Path.home()))
+    sys.path.append(
+        "{}/Developer/ball_e_gui/src/windows".format(pathlib.Path.home()))
 
-import cv2
-import numpy as np
-from PyQt5 import QtGui
-from PyQt5.QtCore import Qt, QThread, pyqtSignal, pyqtSlot
-from PyQt5.QtGui import QPixmap
-from PyQt5.QtWidgets import (QApplication, QHBoxLayout, QLabel, QMainWindow,
-                             QPushButton, QSizePolicy, QVBoxLayout, QWidget)
+    from component_button import GenericButton
+    from component_labels import ProfileLabel
+    from component_toolbar import ToolbarComponent
+    from window_test import TestWindow
 
-import style_constants as sc
-from component_button import GenericButton
-from component_labels import ProfileLabel
-from component_toolbar import ToolbarComponent
-from window_test import TestWindow
+except ImportError:
+    print("Imports failed")
+finally:
+    import pathlib
+
+    import cv2
+    import numpy as np
+    from PyQt5 import QtGui
+    from PyQt5.QtCore import Qt, QThread, pyqtSignal, pyqtSlot
+    from PyQt5.QtGui import QPixmap
+    from PyQt5.QtWidgets import QApplication, QLabel, QVBoxLayout, QWidget
 
 
 class VideoThread(QThread):
