+++ src/screens/screen_training_goalie_profile_selection.py
@@ -1,31 +1,42 @@
-import sys
+try:
+    import pathlib
+    import sys
+    sys.path.append(
+        "{}/Developer/ball_e_gui/src/components".format(pathlib.Path.home()))
+    sys.path.append(
+        "{}/Developer/ball_e_gui/src/helpers".format(pathlib.Path.home()))
+    sys.path.append(
+        "{}/Developer/ball_e_gui/src/windows".format(pathlib.Path.home()))
 
-from PyQt5 import QtWidgets
-from PyQt5.QtCore import Qt
-from PyQt5.QtGui import QFont
-from PyQt5.QtWidgets import (QApplication, QHBoxLayout, QHeaderView, QLabel,
-                             QMainWindow, QPushButton, QSizePolicy,
-                             QTableWidget, QTableWidgetItem, QVBoxLayout,
-                             QWidget)
+    import style_constants as sc
+    from component_button import GenericButton
+    from component_labels import ProfileLabel
+    from component_toolbar import ToolbarComponent
+    from helper_profiler import Profiler
+    from window_test import TestWindow
 
-import style_constants as sc
-from component_button import GenericButton
-from component_labels import ProfileLabel
-from component_toolbar import ToolbarComponent
-from helper_profiler import Profiler
-from window_test import TestWindow
+except ImportError:
+    print("Imports failed")
+finally:
+
+    from PyQt5 import QtWidgets
+    from PyQt5.QtCore import Qt
+    from PyQt5.QtGui import QFont
+    from PyQt5.QtWidgets import (QApplication, QHBoxLayout, QHeaderView,
+                                 QSizePolicy, QTableWidget, QTableWidgetItem,
+                                 QVBoxLayout, QWidget)
 
 
-class TrainingDrillProfileSelectionScreen(QWidget):
-    def __init__(self, selected_goalie_profile=None):
-        super().__init__()
+class TrainingGoalieProfileSelectionScreen(QWidget):
+    def __init__(self, parent=None):
+        super().__init__(parent=parent)
 
-        self.window_title = "Drill Profile Selection"
+        self.window_title = "Goalie Profile Selection"
 
-        self.profiler = Profiler('drill_profiles')
-        self.drill_profiles = self.profiler.get_profiles()
+        self.profiler = Profiler('goalie_profiles')
+        self.goalie_profiles = self.profiler.get_profiles()
 
-        self.selected_drill_profile = None
+        self.selected_goalie_profile = None
 
         # This font will be used for all table related purposes
         self.table_font = QFont()
@@ -33,20 +44,15 @@ class TrainingDrillProfileSelectionScreen(QWidget):
 
         self.screen_layout = QVBoxLayout()
 
-        if selected_goalie_profile is None:
-            self.toolbar = ToolbarComponent(
-                self.window_title, "Back to Training")
-        else:
-            self.toolbar = ToolbarComponent(
-                self.window_title, "Back to Goalie\n Profile Select")
+        self.toolbar = ToolbarComponent(self.window_title, "Back to Training")
 
         self.screen_layout.addWidget(self.toolbar)
 
         screen_layout_row_one = QHBoxLayout()
 
-        self.drill_profile_selection_label = ProfileLabel(
-            "Please Select a Drill Profile to Continue")
-        screen_layout_row_one.addWidget(self.drill_profile_selection_label)
+        self.goalie_profile_selection_label = ProfileLabel(
+            "Please Select a Goalie Profile to Continue")
+        screen_layout_row_one.addWidget(self.goalie_profile_selection_label)
 
         self.next_page_button = GenericButton("Next")
         self.next_page_button.setVisible(False)
@@ -54,12 +60,12 @@ class TrainingDrillProfileSelectionScreen(QWidget):
         self.screen_layout.addLayout(screen_layout_row_one)
 
         self.create_table_header_view(
-            table_title_name="Drill Profiles", header_clicked_action=self.unselect_table_header)
+            table_title_name="Goalie Profiles", header_clicked_action=self.unselect_table_header)
         self.table_header.resizeRowsToContents()
         self.screen_layout.addWidget(self.table_header)
 
         # Create the main table and add to the layout
-        self.create_main_table_view(profile_dict_obj=self.drill_profiles,
+        self.create_main_table_view(profile_dict_obj=self.goalie_profiles,
                                     table_clicked_action=self.choose_main_table_click_action)
         self.main_table_view.resizeRowsToContents()
         self.screen_layout.addWidget(self.main_table_view)
@@ -71,10 +77,10 @@ class TrainingDrillProfileSelectionScreen(QWidget):
         """
         self.screen_layout.removeWidget(self.main_table_view)
 
-        self.drill_profiles = self.profiler.get_profiles()
+        self.goalie_profiles = self.profiler.get_profiles()
 
         # Create the main table and add to the layout
-        self.create_main_table_view(profile_dict_obj=self.drill_profiles,
+        self.create_main_table_view(profile_dict_obj=self.goalie_profiles,
                                     table_clicked_action=self.choose_main_table_click_action)
         self.main_table_view.resizeRowsToContents()
         self.screen_layout.addWidget(self.main_table_view)
@@ -86,7 +92,7 @@ class TrainingDrillProfileSelectionScreen(QWidget):
             table_title_name (str): Table's title
         """
 
-        # Create the table object that acts as the "header" for the main Drill profile table view
+        # Create the table object that acts as the "header" for the main goalie profile table view
         self.table_header = QTableWidget()
         # Do not allow the user to edit the content of the table
         self.table_header.setEditTriggers(
@@ -114,10 +120,10 @@ class TrainingDrillProfileSelectionScreen(QWidget):
 
         # Size the header appropriately
         header_hor_head.setSectionResizeMode(0, QHeaderView.Stretch)
-        # self.table_header.setSizePolicy(
-        #     QSizePolicy.Preferred,
-        #     QSizePolicy.Fixed
-        # )
+        self.table_header.setSizePolicy(
+            QSizePolicy.Preferred,
+            QSizePolicy.Fixed
+        )
         self.table_header.setFixedHeight(
             int(self.table_header.sizeHint().height()/2))
 
@@ -168,28 +174,25 @@ class TrainingDrillProfileSelectionScreen(QWidget):
         main_table_hor_head.setSectionResizeMode(0, QHeaderView.Stretch)
 
     def choose_main_table_click_action(self, item):
-        drill_name = self.main_table_view.item(item.row(), 0).text()
+        goalie_name = self.main_table_view.item(item.row(), 0).text()
 
-        self.drill_profile_selection_label.setText("You have selected: {}".format(
-            drill_name
+        self.goalie_profile_selection_label.setText("You have selected: {}".format(
+            goalie_name
         ))
         self.next_page_button.setVisible(True)
 
-        self.selected_drill_profile = drill_name.replace(' ', '_').lower()
+        self.selected_goalie_profile = goalie_name.replace(' ', '_').lower()
 
     def reset_screen(self):
         # Unselect the currently picked cell
-        self.main_table_view.selectedItems()[0].setSelected(False)
+        self.main_table_view.selectedItems()[
+            0].setSelected(False)
 
-        self.drill_profile_selection_label.setText(
-            "Please Select a Drill Profile to Continue")
+        self.goalie_profile_selection_label.setText(
+            "Please Select a Goalie Profile to Continue")
 
-        self.selected_drill_profile = None
-
-        self.next_page_button.setVisible(False)
-
-    def get_selected_drill_profile(self):
-        return self.selected_drill_profile
+    def get_selected_goalie_profile(self):
+        return self.selected_goalie_profile
 
     def get_window_title(self):
         return self.window_title
@@ -197,7 +200,7 @@ class TrainingDrillProfileSelectionScreen(QWidget):
 
 def main():
     app = QApplication(sys.argv)
-    win = TestWindow(TrainingDrillProfileSelectionScreen())
+    win = TestWindow(TrainingGoalieProfileSelectionScreen())
     win.show()
     sys.exit(app.exec_())
